Move the new nodes to a supernodes package
