+++ src/trufflesom/interpreter/nodes/MessageSendNode.java
@@ -10,6 +10,10 @@ import bd.primitives.nodes.PreevaluatedExpression;
 import trufflesom.interpreter.nodes.ArgumentReadNode.LocalArgumentReadNode;
 import trufflesom.interpreter.nodes.dispatch.GenericDispatchNode;
 import trufflesom.interpreter.nodes.dispatch.UninitializedDispatchNode;
+import trufflesom.interpreter.supernodes.BinaryArgSendNode;
+import trufflesom.interpreter.supernodes.QuatArgSendNode;
+import trufflesom.interpreter.supernodes.TernaryArgSendNode;
+import trufflesom.interpreter.supernodes.UnaryArgSendNode;
 import trufflesom.primitives.Primitives;
 import trufflesom.vm.NotYetImplementedException;
 import trufflesom.vmobjects.SClass;
