+++ src/trufflesom/interpreter/supernodes/UnaryArgSendNode.java
@@ -1,37 +1,30 @@
-package trufflesom.interpreter.nodes;
+package trufflesom.interpreter.supernodes;
 
 import com.oracle.truffle.api.CompilerAsserts;
 import com.oracle.truffle.api.frame.VirtualFrame;
 
+import trufflesom.interpreter.nodes.AbstractMessageSendNode;
 import trufflesom.interpreter.nodes.dispatch.AbstractDispatchNode;
 import trufflesom.interpreter.nodes.dispatch.GenericDispatchNode;
 import trufflesom.vmobjects.SSymbol;
 
 
-public class TernaryArgSendNode extends AbstractMessageSendNode {
+public final class UnaryArgSendNode extends AbstractMessageSendNode {
   private final int                   argIdx;
   private final SSymbol               selector;
   @Child private AbstractDispatchNode dispatchNode;
-  @Child private ExpressionNode       arg1;
-  @Child private ExpressionNode       arg2;
 
-  TernaryArgSendNode(final int argIdx, final ExpressionNode arg1, final ExpressionNode arg2,
-      final SSymbol selector,
+  public UnaryArgSendNode(final int argIdx, final SSymbol selector,
       final AbstractDispatchNode dispatchNode) {
     super(null);
     this.argIdx = argIdx;
     this.selector = selector;
     this.dispatchNode = dispatchNode;
-    this.arg1 = arg1;
-    this.arg2 = arg2;
   }
 
   @Override
   public Object executeGeneric(final VirtualFrame frame) {
-    Object[] arguments = new Object[] {
-        frame.getArguments()[argIdx],
-        arg1.executeGeneric(frame),
-        arg2.executeGeneric(frame)};
+    Object[] arguments = new Object[] {frame.getArguments()[argIdx]};
     return doPreEvaluated(frame, arguments);
   }
 
@@ -48,7 +41,7 @@ public class TernaryArgSendNode extends AbstractMessageSendNode {
 
   @Override
   public int getNumberOfArguments() {
-    return 3;
+    return 1;
   }
 
   @Override
